+++ src/ebfm/coupling/couplers/yacCoupler.py
@@ -13,6 +13,8 @@ from .base import Coupler, Grid, GridDict, CouplingConfig, CouplerErrorCode
 from ebfm.coupling.fields import FieldSet, Field, GenericExchangeType
 from ebfm.coupling.fields import YACField
 
+from ebfm.core.config import Calendar
+
 # from ebfm.geometry import Grid  # TODO: consider introducing a new data structure native to EBFM?
 
 logger = logging.getLogger(__name__)
@@ -33,6 +35,19 @@ class YACCoupler(Coupler[yac.ExchangeType]):
 
         yac_performs_mpi_handshake = not coupling_config.has_group_communicators()
 
+        match coupling_config.time_config.calendar:
+            case Calendar.PROLEPTIC_GREGORIAN:
+                yac.def_calendar(yac.Calendar.PROLEPTIC_GREGORIAN)
+            case Calendar.YEAR_OF_365_DAYS:
+                yac.def_calendar(yac.Calendar.YEAR_OF_365_DAYS)
+            case Calendar.YEAR_OF_360_DAYS:
+                yac.def_calendar(yac.Calendar.YEAR_OF_360_DAYS)
+            case _:
+                raise ValueError(
+                    f"Unsupported calendar '{coupling_config.time_config.calendar}'. Supported calendars are: "
+                    f"{[c.value for c in yac.Calendar]}"
+                )
+
         if yac_performs_mpi_handshake:
             # MPI handshake has not been performed yet, so YAC will perform the splitting
             self.interface = yac.YAC()
@@ -49,6 +64,10 @@ class YACCoupler(Coupler[yac.ExchangeType]):
             coupling_comm = coupling_config.get_group_communicator(yac_group_label)
             logging.debug(f"YAC uses the following comm from EBFM's mpi-handshake: {yac_group_label}: {coupling_comm}.")
             self.interface = yac.YAC(comm=coupling_comm)
+
+        # candidate for abstract function in Coupler base class?
+        # self.initialize_time_frame(coupling_config.start_time, coupling_config.end_time)
+        self.interface.def_datetime(coupling_config.start_time, coupling_config.end_time)
 
         if coupling_config.coupler_config:
             self.interface.read_config_yaml(str(coupling_config.coupler_config))